Removing warnings for the application. Except one warning which needs more work.

+++ MapView/Map/RMTileSourcesContainer.m
@@ -40,8 +40,8 @@
 
     RMSphericalTrapezium _latitudeLongitudeBoundingBox;
 
-    float _minZoom, _maxZoom;
-    int _tileSideLength;
+    CGFloat _minZoom, _maxZoom;
+    NSUInteger _tileSideLength;
 }
 
 - (id)init
